+++ src/alhambra_mixes/logging.py
@@ -1,0 +1,3 @@
+import logging
+
+log = logging.getLogger("alhambra")
